docker/paddleocr/requirements.txt
paddlepaddle==2.5.2
paddleocr==2.7.0
Pillow>=9.0.0
numpy>=1.19.0
pdf2image>=1.16.0
fastapi>=0.104.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-json-logger>=2.0.0